requirements.txt
langchain>=0.2.0
langchain-community
langchain-google-genai
google-generativeai
chromadb
python-dotenv
langchain-google-genai
faiss-cpu
docx2txt